refactor(krt): the fake resolver of the tests lives in one module
The tests of the fold and the tests of the run loop each held a copy of
FakeResolver, of the named() helper, and of the impl that gives the fake to
the Resolver trait. The two copies encode one contract: the first ask of an
address starts the lookup, and every ask after it reads the answer. Two
copies of a contract drift, and a correction to one copy leaves the other
asserting the old behavior. The fake now lives in crate::testing, beside the
fixtures that more than one module already shares, and both test modules
import it.

The move touches test code alone, so it needs no red test: the proof is that
the suite passes identically before and after. The counts before the change
were 335, 12, and 18 passed, and the counts after it are 335, 12, and 18
passed, with no failure on either side. A changed count would mean a test
moved or went away, which this change does not do.

TARGET_NAME stays where it is in each module. The two definitions carry
different values, "example.com" in names.rs and "target.example.com" in
run.rs, so no merge of them is correct. The RUN constants differ the same
way, so the namer() and a_namer() helpers that read them stay in place.
