+++ src/krt/src/testing.rs
@@ -6,12 +6,22 @@
 //! and each test names only the part it cares about: the TTLs that the round
 //! probed, and the hops that answered.
 //!
+//! The module also holds the fake resolver that a test of the fold and a test
+//! of the run loop both program. The fake stands for one contract: the first
+//! ask of an address starts the lookup, and every ask after it reads the
+//! answer. One copy of the fake holds one statement of that contract, so a
+//! correction to the contract reaches every test of it.
+//!
 //! This module compiles under `cfg(test)` alone, so nothing it holds reaches
 //! the binary.
 
+use crate::names::{Lookup, Resolver};
 use crate::record::{Hop, RoundRecord, RunId, TtlRange};
 use chrono::{DateTime, Utc};
+use std::cell::{Cell, RefCell};
+use std::collections::{HashMap, VecDeque};
 use std::net::IpAddr;
+use std::rc::Rc;
 
 /// The identifier of the run that every test round belongs to.
 const RUN: &str = "2026-08-18T12:00:00.000Z";
@@ -64,5 +74,68 @@ pub(crate) fn round(first: u8, last: u8, hops: &[(u8, &str, f64)]) -> RoundRecor
                 icmp: TIME_EXCEEDED.to_owned(),
             })
             .collect(),
+    }
+}
+
+/// The answer of a lookup that finished with a name.
+pub(crate) fn named(host: &str) -> Lookup {
+    Lookup::Named(host.to_owned())
+}
+
+/// A resolver that a test programs: one answer for each ask of one address,
+/// and the last answer of the list for every ask after the list runs out.
+///
+/// An address that the test named no answer for answers `Nameless`.
+///
+/// The count and the answers sit behind a `Cell` and a `RefCell`, because
+/// [`Resolver::lookup`] takes the resolver by reference. The fake stays on
+/// one thread.
+pub(crate) struct FakeResolver {
+    /// The answers that each address holds, the next answer first.
+    answers: RefCell<HashMap<IpAddr, VecDeque<Lookup>>>,
+    /// The number of asks that the resolver took.
+    asks: Cell<usize>,
+}
+
+impl FakeResolver {
+    /// A resolver that answers each address with the answers of its list.
+    pub(crate) fn new(answers: &[(&str, &[Lookup])]) -> Rc<Self> {
+        Rc::new(Self {
+            answers: RefCell::new(
+                answers
+                    .iter()
+                    .map(|(addr, list)| (address(addr), list.iter().cloned().collect()))
+                    .collect(),
+            ),
+            asks: Cell::new(0),
+        })
+    }
+
+    /// The number of asks that the resolver took.
+    pub(crate) fn asks(&self) -> usize {
+        self.asks.get()
+    }
+
+    /// The answer of one ask, and one step along the list of that address.
+    fn answer(&self, addr: IpAddr) -> Lookup {
+        self.asks.set(self.asks.get() + 1);
+        let mut answers = self.answers.borrow_mut();
+        let Some(queue) = answers.get_mut(&addr) else {
+            return Lookup::Nameless;
+        };
+        // The last answer of a list stands for every ask after it, so the
+        // list keeps that answer in the place of a step.
+        let answer = if queue.len() > 1 {
+            queue.pop_front()
+        } else {
+            queue.front().cloned()
+        };
+        answer.unwrap_or(Lookup::Nameless)
+    }
+}
+
+impl Resolver for Rc<FakeResolver> {
+    fn lookup(&self, addr: IpAddr) -> Lookup {
+        self.answer(addr)
     }
 }
